Rename TreeView component to Tree for consistency
The component has been simplified and renamed from TreeView to Tree
to better align with other component naming conventions. This change
includes renaming related files, updating component references, and
refreshing the documentation to reflect the new naming scheme.

Icons can now be specified as in other components as well.

Removed the Navigation component because I hated it and will remake it
with something easier to use.